backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
pinecone-client==3.0.0
celery==5.3.4
websockets==12.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
google-cloud-aiplatform==1.36.4
google-generativeai==0.3.2
supabase==2.0.2
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
aiofiles==23.2.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1